create custom hook for the displayedOfferings settings
